Move isCurrentlyPausedAtBreakpoint
