+++ public/js/components/Scopes.css
@@ -10,7 +10,6 @@
   margin-bottom: 20px;
 }
 
-
 .scopes-pane ul.scope-property-list {
   list-style: none;
   padding: 0;
